chore: rename project to opencode-a2a-server #170
