+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
         run: bash ./scripts/lint.sh
 
       - name: Run mypy
-        run: uv run mypy src/opencode_a2a_serve
+        run: uv run mypy src/opencode_a2a_server
 
       - name: Run pytest
         run: uv run pytest
