+++ SECURITY.md
@@ -11,7 +11,7 @@ not fully isolate upstream model credentials from OpenCode runtime behavior.
 
 - `A2A_BEARER_TOKEN` protects access to the A2A surface, but it is not a
   tenant-isolation boundary inside one deployed instance.
-- One `OpenCode + opencode-a2a-serve` instance pair is treated as a
+- One `OpenCode + opencode-a2a-server` instance pair is treated as a
   single-tenant trust boundary by design.
 - Within one instance, consumers share the same underlying OpenCode
   workspace/environment by default.
