+++ src/opencode_a2a_server/extension_contracts.py
