+++ src/opencode_a2a_server/jsonrpc_ext.py
