+++ src/opencode_a2a_server/text_parts.py
