+++ tests/helpers.py
@@ -7,8 +7,8 @@ from a2a.server.agent_execution import RequestContext
 from a2a.server.context import ServerCallContext
 from a2a.types import Message, MessageSendParams, Role, TextPart
 
-from opencode_a2a_serve.config import Settings
-from opencode_a2a_serve.opencode_client import OpencodeMessage
+from opencode_a2a_server.config import Settings
+from opencode_a2a_server.opencode_client import OpencodeMessage
 
 
 def make_settings(**overrides: Any) -> Settings:
