+++ tests/test_directory_validation.py
@@ -4,8 +4,8 @@ from unittest.mock import AsyncMock
 import pytest
 from a2a.server.events.event_queue import EventQueue
 
-from opencode_a2a_serve.agent import OpencodeAgentExecutor
-from opencode_a2a_serve.opencode_client import OpencodeClient
+from opencode_a2a_server.agent import OpencodeAgentExecutor
+from opencode_a2a_server.opencode_client import OpencodeClient
 from tests.helpers import make_request_context_mock, make_settings
