+++ tests/test_session_ownership.py
@@ -4,8 +4,8 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 from a2a.server.events.event_queue import EventQueue
 
-from opencode_a2a_serve.agent import OpencodeAgentExecutor, _TTLCache
-from opencode_a2a_serve.opencode_client import OpencodeClient
+from opencode_a2a_server.agent import OpencodeAgentExecutor, _TTLCache
+from opencode_a2a_server.opencode_client import OpencodeClient
 from tests.helpers import configure_mock_client_runtime, make_request_context_mock
