+++ tests/test_transport_contract.py
@@ -4,7 +4,7 @@ import httpx
 import pytest
 from a2a.types import TransportProtocol
 
-from opencode_a2a_serve.app import _normalize_log_level, build_agent_card, create_app
+from opencode_a2a_server.app import _normalize_log_level, build_agent_card, create_app
 from tests.helpers import DummyChatOpencodeClient, make_settings
 
 
@@ -67,7 +67,7 @@ def test_openapi_jsonrpc_examples_include_core_message_methods() -> None:
 
 @pytest.mark.asyncio
 async def test_dual_stack_send_accepts_transport_native_payloads(monkeypatch) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(make_settings(a2a_bearer_token="test-token"))
@@ -105,7 +105,7 @@ async def test_dual_stack_send_accepts_transport_native_payloads(monkeypatch) ->
 
 @pytest.mark.asyncio
 async def test_dual_stack_send_rejects_cross_transport_payload_shapes(monkeypatch) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(make_settings(a2a_bearer_token="test-token"))
@@ -194,14 +194,14 @@ def _jsonrpc_message_send_payload(text: str) -> dict:
 
 @pytest.mark.asyncio
 async def test_log_payloads_keeps_body_for_rest_handler(monkeypatch, caplog) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(make_settings(a2a_bearer_token="test-token", a2a_log_payloads=True))
     transport = httpx.ASGITransport(app=app)
     headers = {"Authorization": "Bearer test-token"}
 
-    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_serve.app"):
+    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_server.app"):
         async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
             resp = await client.post(
                 "/v1/message:send",
@@ -218,14 +218,14 @@ async def test_log_payloads_keeps_body_for_rest_handler(monkeypatch, caplog) -> 
 
 @pytest.mark.asyncio
 async def test_log_payloads_streaming_response_path(monkeypatch, caplog) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(make_settings(a2a_bearer_token="test-token", a2a_log_payloads=True))
     transport = httpx.ASGITransport(app=app)
     headers = {"Authorization": "Bearer test-token"}
 
-    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_serve.app"):
+    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_server.app"):
         async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
             async with client.stream(
                 "POST", "/v1/message:stream", headers=headers, json=_rest_message_payload()
@@ -245,7 +245,7 @@ async def test_log_payloads_streaming_response_path(monkeypatch, caplog) -> None
 
 @pytest.mark.asyncio
 async def test_log_payloads_omits_non_json_request_body(monkeypatch, caplog) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(make_settings(a2a_bearer_token="test-token", a2a_log_payloads=True))
@@ -255,7 +255,7 @@ async def test_log_payloads_omits_non_json_request_body(monkeypatch, caplog) -> 
         "Content-Type": "application/octet-stream",
     }
 
-    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_serve.app"):
+    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_server.app"):
         async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
             resp = await client.post("/", headers=headers, content=b"\x00\x01\x02\x03")
             assert resp.status_code < 500
@@ -269,7 +269,7 @@ async def test_log_payloads_omits_non_json_request_body(monkeypatch, caplog) -> 
 
 @pytest.mark.asyncio
 async def test_log_payloads_omits_text_plain_request_body(monkeypatch, caplog) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(make_settings(a2a_bearer_token="test-token", a2a_log_payloads=True))
@@ -283,7 +283,7 @@ async def test_log_payloads_omits_text_plain_request_body(monkeypatch, caplog) -
         '{"messageId":"m","role":"user","parts":[{"kind":"text","text":"secret"}]}}}'
     )
 
-    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_serve.app"):
+    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_server.app"):
         async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
             resp = await client.post("/", headers=headers, content=body)
             assert resp.status_code < 500
@@ -297,7 +297,7 @@ async def test_log_payloads_omits_text_plain_request_body(monkeypatch, caplog) -
 
 @pytest.mark.asyncio
 async def test_log_payloads_omits_when_content_length_missing(monkeypatch, caplog) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(
@@ -320,7 +320,7 @@ async def test_log_payloads_omits_when_content_length_missing(monkeypatch, caplo
     async def _body_stream():
         yield body
 
-    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_serve.app"):
+    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_server.app"):
         async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
             resp = await client.post(
                 "/",
@@ -342,7 +342,7 @@ async def test_log_payloads_omits_when_content_length_missing(monkeypatch, caplo
 
 @pytest.mark.asyncio
 async def test_log_payloads_omits_oversized_request_body(monkeypatch, caplog) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyChatOpencodeClient)
     app = app_module.create_app(
@@ -356,7 +356,7 @@ async def test_log_payloads_omits_oversized_request_body(monkeypatch, caplog) ->
     headers = {"Authorization": "Bearer test-token"}
     oversized_text = "x" * 512
 
-    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_serve.app"):
+    with caplog.at_level(logging.DEBUG, logger="opencode_a2a_server.app"):
         async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
             resp = await client.post(
                 "/",
@@ -373,7 +373,7 @@ async def test_log_payloads_omits_oversized_request_body(monkeypatch, caplog) ->
 
 
 def test_create_app_propagates_cancel_abort_timeout(monkeypatch) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     captured: dict[str, float | bool | int] = {}
 
@@ -429,7 +429,7 @@ def test_create_app_propagates_cancel_abort_timeout(monkeypatch) -> None:
 
 
 def test_create_app_requires_control_guard_hooks(monkeypatch) -> None:
-    import opencode_a2a_serve.app as app_module
+    import opencode_a2a_server.app as app_module
 
     class _BrokenExecutor:
         def __init__(
